requirements.txt
fastapi
uvicorn
python-dotenv
aiohttp
numpy
pydantic
beautifulsoup4 
html2text
tqdm
markdown
python-multipart
setuptools
playwright
markdownify